Implemented the addition of a new note, and categories

+++ src/components/DashItem.jsx
@@ -21,7 +21,7 @@ const CustomButton = styled(Button)(({ theme }) => ({
 }));
 
 
-const DashItem = ({ title, board, categoryList }) => {
+const DashItem = ({ title, board, categoryList, totalCount }) => {
   const dispatch = useDispatch()
   const [openModal, setOpenModal] = React.useState(false);
   const { loading } = useSelector(({boards}) => boards)
@@ -33,7 +33,9 @@ const DashItem = ({ title, board, categoryList }) => {
   }
 
   const deleteTask = (id, listId) => {
-    dispatch(deleteTaskThunk(id, listId))
+    if (window.confirm("Do you really want to delete note?")) {
+      dispatch(deleteTaskThunk(id, listId))
+    }
   }
 
   const handleOpenModal = () => setOpenModal(true)
@@ -54,10 +56,10 @@ const DashItem = ({ title, board, categoryList }) => {
         </Box>
         <CustomButton fullWidth onClick={handleOpenModal}><Typography variant={'h5'} sx={{ fontSize: '30px' }}>+</Typography></CustomButton>
         {
-          !loading ? board.tasks.map((obj) => <Task {...obj} categoryList={categoryList} deleteTask={deleteTask}/>)  : <LoadingBoards/>
+          !loading ? board.tasks.map((obj) => <Task {...obj}  categoryList={categoryList} deleteTask={deleteTask}/>)  : <LoadingBoards/>
         }
       </CustomPaper>
-      <AddNewTask listId={listId} setOpenModal={setOpenModal} open={openModal} addTask={addTask} handleCloseModal={handleCloseModal}/>
+      <AddNewTask listId={listId} totalCount={totalCount} setOpenModal={setOpenModal} open={openModal} addTask={addTask} handleCloseModal={handleCloseModal}/>
     </Grid>
   );
 };
